Вся shared elements анимация удалена
