+++ app/src/main/java/pro/audasviator/yamobilizationapp/ArtistListFragment.java
@@ -1,32 +1,22 @@
 package pro.audasviator.yamobilizationapp;
 
-import android.annotation.TargetApi;
 import android.content.Context;
-import android.content.Intent;
 import android.content.res.TypedArray;
 import android.graphics.Canvas;
 import android.graphics.drawable.Drawable;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.AsyncTask;
-import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v4.app.SharedElementCallback;
-import android.support.v4.view.ViewCompat;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.transition.ChangeBounds;
-import android.transition.ChangeImageTransform;
-import android.transition.ChangeTransform;
-import android.transition.TransitionSet;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.ViewTreeObserver;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -36,73 +26,25 @@ import com.squareup.picasso.Picasso;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 public class ArtistListFragment extends Fragment {
-    public static final String EXTRA_STARTING_POSITION = "startingPosition";
-    public static final String EXTRA_CURRENT_POSITION = "currentPosition";
-
-    private static final int REQUEST_CODE = 42;
-
     private RecyclerView mRecyclerView;
     private ArtistLab mArtistLab;
     private List<Artist> mArtistList;
     private SwipeRefreshLayout mRefreshLayout;
+    private int mScrollTo; // Костыль :(
 
-    private boolean mIsDetailsActivityStarted; // Вдруг пользователь просто свернул активти с этим фрагментом?
-    private Bundle mTmpReenterState; // Информация о shared elements (id|tag вьюхи, показывающей обложку)
-
-    // Вызывается перед началом анимации в другое активити
-    private final SharedElementCallback mCallback = new SharedElementCallback() {
-        @Override
-        public void onMapSharedElements(List<String> names, Map<String, View> sharedElements) {
-            if (mTmpReenterState != null) {
-                int startingPosition = mTmpReenterState.getInt(EXTRA_STARTING_POSITION); // Откуда начали скоролить ViewPager
-                int currentPosition = mTmpReenterState.getInt(EXTRA_CURRENT_POSITION); // Куда доскролили
-                if (startingPosition != currentPosition) {
-                    // Добавляем новую обложку из ViewPager`а в shared elements
-                    String newTransitionName = String.valueOf(mArtistList.get(currentPosition).getId());
-                    View newSharedElement = mRecyclerView.findViewWithTag(newTransitionName);
-                    if (newSharedElement != null) {
-                        names.clear();
-                        names.add(newTransitionName);
-                        sharedElements.clear();
-                        sharedElements.put(newTransitionName, newSharedElement);
-                    }
-                }
-
-                mTmpReenterState = null;
-            }
-        }
-    };
 
     public static ArtistListFragment newInstance() {
         return new ArtistListFragment();
-    }
-
-    // Взывается DetailFragment, дабы передать откуда и куда доскролился ViewPager
-    public static Intent makeIntentForAnimation(int currentPosition, int startingPosition) {
-        Intent data = new Intent();
-        data.putExtra(EXTRA_STARTING_POSITION, startingPosition);
-        data.putExtra(EXTRA_CURRENT_POSITION, currentPosition);
-        return data;
     }
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        //AppCompatActivity activity = (AppCompatActivity) getActivity();
-        //activity.setExitSharedElementCallback(mCallback);
-
         mArtistLab = ArtistLab.get(getActivity());
         mArtistList = mArtistLab.getArtists();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        mIsDetailsActivityStarted = false;
     }
 
     @Nullable
@@ -123,6 +65,9 @@ public class ArtistListFragment extends Fragment {
         mRecyclerView.setAdapter(new ArtistAdapter(mArtistList));
         mRecyclerView.addItemDecoration(new amazingDividerItemDecoration(getContext()));
 
+        // Скроллим RecyclerView до того же места, что и ViewPager
+        mRecyclerView.getLayoutManager().smoothScrollToPosition(mRecyclerView, null, mScrollTo);
+
         // Обновим за пользователя для первого раза
         if (mArtistList.size() == 0) {
             new FetchArtistTask().execute(getContext());
@@ -131,39 +76,8 @@ public class ArtistListFragment extends Fragment {
         return view;
     }
 
-    // Подробности в ArtistDetailActivity
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        mTmpReenterState = new Bundle(data.getExtras());
-        int startingPosition = mTmpReenterState.getInt(EXTRA_STARTING_POSITION);
-        int currentPosition = mTmpReenterState.getInt(EXTRA_CURRENT_POSITION);
-        if (startingPosition != currentPosition) {
-            mRecyclerView.scrollToPosition(currentPosition);
-        }
-    }
-
-    public void onActivityReenter(Intent data) {
-        mTmpReenterState = new Bundle(data.getExtras());
-        int startingPosition = mTmpReenterState.getInt(EXTRA_STARTING_POSITION);
-        int currentPosition = mTmpReenterState.getInt(EXTRA_CURRENT_POSITION);
-        if (startingPosition != currentPosition) {
-            mRecyclerView.scrollToPosition(currentPosition);
-        }
-
-        getActivity().supportPostponeEnterTransition();
-        mRecyclerView.getViewTreeObserver().addOnPreDrawListener(new ViewTreeObserver.OnPreDrawListener() {
-            @Override
-            public boolean onPreDraw() {
-                mRecyclerView.getViewTreeObserver().removeOnPreDrawListener(this);
-
-                // Сначала должны забиндиться вьюхи (дабы понять, какую обложку двигать)
-                mRecyclerView.requestLayout();
-
-                // А уже потом должна сработать анимация
-                getActivity().supportStartPostponedEnterTransition();
-                return true;
-            }
-        });
+    public void scrollTo(int position) {
+        mScrollTo = position;
     }
 
     private class ArtistHolder extends RecyclerView.ViewHolder {
@@ -184,38 +98,13 @@ public class ArtistListFragment extends Fragment {
             itemView.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
-                    Intent intent = ArtistDetailActivity.newIntent(getContext(), mPosition);
-                    intent.putExtra(EXTRA_STARTING_POSITION, mPosition);
 
-                    if (!mIsDetailsActivityStarted) { // Дабы не тыкали дважды
-                        mIsDetailsActivityStarted = true;
-                        //ActivityOptionsCompat options = ActivityOptionsCompat.makeSceneTransitionAnimation(getActivity(), mCoverImageView, ViewCompat.getTransitionName(mCoverImageView));
-                        //startActivityForResult(intent, REQUEST_CODE, options.toBundle());
-
-                        ArtistViewPagerFragment artistViewPagerFragment = ArtistViewPagerFragment.newInstance(mPosition);
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-
-                            @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-                            class DetailsTransition extends TransitionSet {
-                                public DetailsTransition() {
-                                    setOrdering(ORDERING_TOGETHER);
-                                    addTransition(new ChangeBounds()).
-                                            addTransition(new ChangeTransform()).
-                                            addTransition(new ChangeImageTransform());
-                                }
-                            }
-
-                            artistViewPagerFragment.setSharedElementEnterTransition(new DetailsTransition());
-                            artistViewPagerFragment.setSharedElementReturnTransition(new DetailsTransition());
-                        }
-
-                        getFragmentManager()
-                                .beginTransaction()
-                                .replace(R.id.fragment_list_container, artistViewPagerFragment)
-                                .addToBackStack(null)
-                                .addSharedElement(mCoverImageView, ViewCompat.getTransitionName(mCoverImageView))
-                                .commit();
-                    }
+                    ArtistViewPagerFragment artistViewPagerFragment = ArtistViewPagerFragment.newInstance(mPosition);
+                    getFragmentManager()
+                            .beginTransaction()
+                            .replace(R.id.fragment_list_container, artistViewPagerFragment)
+                            .addToBackStack(null)
+                            .commit();
                 }
             });
         }
@@ -234,11 +123,6 @@ public class ArtistListFragment extends Fragment {
             mCountTextView.setText(count);
 
             mPosition = position;
-
-            // Используем id в качестве имени для анимации и тега (для поиска вьюхи)
-            String tagAndName = String.valueOf(artist.getId());
-            ViewCompat.setTransitionName(mCoverImageView, tagAndName);
-            mCoverImageView.setTag(tagAndName);
 
             // Кушает кучу памяти (1/7th heap size)
             Picasso.with(getContext()).load(coverUrl).placeholder(R.drawable.the_place_holder).into(mCoverImageView);
